requirements.txt
bs4==0.0.2
numpy==1.23.1
openai==1.35.13
pandas==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.1
scipy==1.13.1
scikit-learn==1.1.2
streamlit==1.12.0
